feat: 알림 조회시 필요한 API,  알림 읽음 상태로 변경 API 개발

+++ src/main/java/pheonix/classconnect/backend/com/common/model/PageRequest.java
@@ -8,24 +8,24 @@ import lombok.Setter;
 @Getter @Setter
 @NoArgsConstructor
 public class PageRequest {
-    private Integer cursorId = 0;
+    private Long cursorId = 0L;
     private Integer page = 1;
     private Integer size = 15;
 
-    public PageRequest(Integer cursorId, Integer page, Integer size){
+    public PageRequest(Long cursorId, Integer page, Integer size){
         this.cursorId = cursorId;
         this.page = page > 1 ? page : 1;
         this.size = size > 0 ? size : 15;
     }
 
     public PageRequest(Integer page, Integer size){
-        this.cursorId = 0;
+        this.cursorId = 0L;
         this.page = page;
         this.size = size;
     }
 
     public PageRequest(Integer page){
-        this.cursorId = 0;
+        this.cursorId = 0L;
         this.page = page;
         this.size = 15;
     }
